Update patch set 1
Patch Set 1:

> Sorry, I missed this comment. Are you on the most recent Buck
 > master on rabbitmq plugin?

No, I'm using the version that we have on Gerrit stable-2.10.  If I use a different version it has to update and rebuild itself every time I switch between building the plugin and building Gerrit.

This is something I wanted to discuss, actually.  Should we have stable-2.x branches on the bucklets git with buck versions matching the versions in Gerrit?

Anyway, for now I'll try building the plugin with the latest version of buck.

Patch-set: 1