Kill PLUGIN_VERSION; always generate version from git describe
By using the version specified in PLUGIN_VERSION, the version of
the plugin will always be the same regardless of which commit it
was actually built from. This makes it extremely difficult to know
which version is installed in a Gerrit server.

Remove support for PLUGIN_VERSION, and always generate the version
from git describe. This way, the sha1 of the commit is included
in the version number.

Keep the previous behaviour of defaulting to "1.0" if the plugin
is not being built from git version controlled source.

Change-Id: I71d18a5054f4ec7a4614adf702d25b209464ac83