Update patch set 2
Patch Set 2:

I think this change will break glitter and other projects that use java_doc bucklets. May be mention this in the commit message or even upload a fix for those projects? This series is really needed if we want (and we do want) to share similar build implementation for in tree and standalone buck build modes.

Patch-set: 2