Install Saxon jar to local maven repository
So far, every time a sonar analysis was done, the Saxon library was
downloaded and stored to a temporal folder. That was inconvenient when
trying to run analysis and no internet connection was available.

Use maven-dependency-plugin:copy so than Saxon jar cached in local maven
repository and available even when off line.

Change-Id: Ib2bd39c3275814fe95e2626b37b572206b85f870