Install Saxon library jar to local maven repository
So far, every time a sonar analysis was done, the Saxon library was
downloaded and stored to a temporal folder. That was inconvenient when
trying to run analysis and no internet connection was available.

Download and install the Saxon jar library to the local maven repository
so that it is available even when off line. Custom repository folder
location is supported.

Change-Id: Ib2bd39c3275814fe95e2626b37b572206b85f870